Move `ConnectivityWindowManager` inside `ConnectivityManager`
Improve Array+Extension docs.
Add `Configuration` to `ConnectivityManager+WindowManager`
